requirements.txt
pytest
python-dotenv
python-binance